+++ apps/api/src/services/data-provider/data-enhancer/yahoo-finance/yahoo-finance.service.ts
@@ -20,7 +20,7 @@ import {
 } from '@prisma/client';
 import { isISIN } from 'class-validator';
 import YahooFinance from 'yahoo-finance2';
-import type { Price } from 'yahoo-finance2/esm/src/modules/quoteSummary-iface';
+import type { Price } from 'yahoo-finance2/modules/quoteSummary-iface';
 
 @Injectable()
 export class YahooFinanceDataEnhancerService implements DataEnhancerInterface {