+++ apps/client/src/app/components/user-account-access/create-or-update-access-dialog/interfaces/interfaces.ts
@@ -1,6 +1,5 @@
 import { Access } from '@ghostfolio/common/interfaces';
 
 export interface CreateOrUpdateAccessDialogParams {
-  // TODO: Include the scopes once the dialog allows to configure them
-  access?: Omit<Access, 'scopes'>;
+  access?: Access;
 }