+++ libs/common/src/lib/interfaces/responses/data-provider-ghostfolio-status-response.interface.ts
@@ -3,5 +3,6 @@ import { UserWithSettings } from '@ghostfolio/common/types';
 export interface DataProviderGhostfolioStatusResponse {
   dailyRequests: number;
   dailyRequestsMax: number;
+  isWithinSetupPeriod: boolean;
   subscription: UserWithSettings['subscription'];
 }